🚧️(Qin): refactor type constraints in Qin class
Introduced a macro, `FORWARD_CONSTRAINT`, in `Qin.h` to remove redundancies in type checking. This results in cleaner and more maintainable code by removing duplicated checks to the type of `V`. These type checks are now represented in the `FORWARD_CONSTRAINT` macro. Furthermore, this change avoids potential related bugs in future if any changes are made pertaining to these conditions. A static cast to `NoneCVT` is added in assignment to `tmp_val` in `set` method to ensure proper conversion.

Signed-off-by: xinbingnan <xinbingnan@xiaomi.com>