🚧(Yi): refactor Qin and create Yi classes to support different input/output types
The class Qin was refactored into a new class Yi with two template parameters INPUT_TYPE and OUTPUT_TYPE, allowing for more flexibility. A utility function 'convert' was added to convert between input and output types if able. A 'set_raw' function was added in Yi to handle raw input. The existing Qin class was redefined to extend from Yi with same types for input and output parameters. This change will allow for more versatile use of these classes in handling different types of input and outputs. Also, the example file was updated to demonstrate the usage of these changes.

Signed-off-by: Benign X <1341398182@qq.com>